Feature/order shop (#102)
* refactor name order and orderGroup

* refactor /user/order-group api and add user/order api

* refactor artist api
