+++ artiexh-api/src/main/java/com/artiexh/api/service/UserService.java
@@ -7,10 +7,18 @@ import com.artiexh.model.rest.user.UserOrderGroupResponse;
 import com.artiexh.model.rest.user.UserOrderGroupResponsePage;
 import com.artiexh.model.rest.user.UserOrderResponse;
 import com.artiexh.model.rest.user.UserOrderResponsePage;
+import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
 
 public interface UserService {
-	UserOrderGroupResponse getOrderById(Long id, Long userId);
-	PageResponse<UserOrderGroupResponsePage> getOrderInPage(Specification<OrderGroupEntity> specification, Pageable pageable);
+
+	UserOrderGroupResponse getOrderGroupById(Long id, Long userId);
+
+	PageResponse<UserOrderGroupResponsePage> getOrderGroupInPage(Specification<OrderGroupEntity> specification, Pageable pageable);
+
+	Page<UserOrderResponsePage> getOrderInPage(Specification<OrderEntity> specification, Pageable pageable);
+
+	UserOrderResponse getOrderById(Long id, Long userId);
+
 }
