+++ artiexh-api/src/main/java/com/artiexh/api/service/impl/UserServiceImpl.java
@@ -1,8 +1,10 @@
 package com.artiexh.api.service.impl;
 
 import com.artiexh.api.exception.ErrorCode;
+import com.artiexh.api.service.OrderGroupService;
 import com.artiexh.api.service.OrderService;
 import com.artiexh.api.service.UserService;
+import com.artiexh.data.jpa.entity.OrderEntity;
 import com.artiexh.data.jpa.entity.OrderGroupEntity;
 import com.artiexh.model.domain.OrderGroup;
 import com.artiexh.model.mapper.OrderGroupMapper;
@@ -10,6 +12,8 @@ import com.artiexh.model.mapper.OrderMapper;
 import com.artiexh.model.rest.PageResponse;
 import com.artiexh.model.rest.user.UserOrderGroupResponse;
 import com.artiexh.model.rest.user.UserOrderGroupResponsePage;
+import com.artiexh.model.rest.user.UserOrderResponse;
+import com.artiexh.model.rest.user.UserOrderResponsePage;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -20,12 +24,13 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public class UserServiceImpl implements UserService {
 	private final OrderService orderService;
+	private final OrderGroupService orderGroupService;
 	private final OrderMapper orderMapper;
 	private final OrderGroupMapper orderGroupMapper;
 
 	@Override
-	public UserOrderGroupResponse getOrderById(Long id, Long userId) {
-		OrderGroup order = orderService.getById(id);
+	public UserOrderGroupResponse getOrderGroupById(Long id, Long userId) {
+		OrderGroup order = orderGroupService.getById(id);
 		if (!order.getUser().getId().equals(userId)) {
 			throw new IllegalArgumentException(ErrorCode.ORDER_IS_INVALID.getMessage());
 		}
@@ -33,8 +38,20 @@ public class UserServiceImpl implements UserService {
 	}
 
 	@Override
-	public PageResponse<UserOrderGroupResponsePage> getOrderInPage(Specification<OrderGroupEntity> specification, Pageable pageable) {
-		Page<OrderGroup> orderPage = orderService.getInPage(specification, pageable);
+	public PageResponse<UserOrderGroupResponsePage> getOrderGroupInPage(Specification<OrderGroupEntity> specification, Pageable pageable) {
+		Page<OrderGroup> orderPage = orderGroupService.getInPage(specification, pageable);
 		return new PageResponse<>(orderPage.map(orderGroupMapper::domainToUserResponsePage));
+	}
+
+	@Override
+	public Page<UserOrderResponsePage> getOrderInPage(Specification<OrderEntity> specification, Pageable pageable) {
+		return orderService.getOrderInPage(specification, pageable)
+			.map(orderMapper::domainToUserResponsePage);
+	}
+
+	@Override
+	public UserOrderResponse getOrderById(Long id, Long userId) {
+		var userOrder = orderService.getOrderByIdAndUserId(id, userId);
+		return orderMapper.domainToUserResponse(userOrder);
 	}
 }
