+++ lib_artiexh-data-jpa/src/main/java/com/artiexh/data/jpa/entity/OrderEntity.java
@@ -28,10 +28,6 @@ public class OrderEntity extends BaseAuditEntity {
 	@Column(name = "id", nullable = false)
 	private Long id;
 
-//	@ManyToOne(fetch = FetchType.LAZY, optional = false)
-//	@JoinColumn(name = "user_id", nullable = false)
-//	private UserEntity user;
-
 	@ManyToOne(fetch = FetchType.LAZY, optional = false)
 	@JoinColumn(name = "shop_id", nullable = false)
 	private ArtistEntity shop;
@@ -40,15 +36,8 @@ public class OrderEntity extends BaseAuditEntity {
 	@JoinColumn(name = "order_group_id", insertable = false, updatable = false)
 	private OrderGroupEntity orderGroup;
 
-//	@ManyToOne(fetch = FetchType.LAZY, optional = false)
-//	@JoinColumn(name = "shipping_address_id", nullable = false)
-//	private UserAddressEntity shippingAddress;
-
 	@Column(name = "note")
 	private String note;
-
-	@Column(name = "payment_method", nullable = false)
-	private Byte paymentMethod;
 
 	@Column(name = "status", nullable = false)
 	private Byte status;
@@ -56,10 +45,6 @@ public class OrderEntity extends BaseAuditEntity {
 	@OneToMany
 	@JoinColumn(name = "order_id")
 	private Set<OrderDetailEntity> orderDetails = new LinkedHashSet<>();
-
-//	@OneToMany
-//	@JoinColumn(name = "order_id")
-//	private Set<OrderTransactionEntity> orderTransaction = new LinkedHashSet<>();
 
 	@OneToMany
 	@JoinColumn(name = "order_id")
