+++ lib_artiexh-model/src/main/java/com/artiexh/model/rest/order/request/CheckoutShop.java
@@ -19,6 +19,7 @@ public class CheckoutShop {
 
 	private String note;
 
+	@NotNull
 	private BigDecimal shippingFee;
 
 	@NotEmpty
