+++ lib_artiexh-model/src/main/java/com/artiexh/model/rest/order/request/OrderGroupPageFilter.java
@@ -1,10 +1,10 @@
 package com.artiexh.model.rest.order.request;
 
 import com.artiexh.data.jpa.entity.OrderEntity;
+import com.artiexh.data.jpa.entity.OrderGroupEntity;
 import com.artiexh.model.domain.OrderStatus;
-import jakarta.persistence.criteria.Path;
+import jakarta.persistence.criteria.Join;
 import jakarta.persistence.criteria.Predicate;
-import jakarta.persistence.criteria.Root;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -14,41 +14,32 @@ import org.springframework.data.jpa.domain.Specification;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.function.Function;
 
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-public class OrderPageFilter {
+public class OrderGroupPageFilter {
+
 	private OrderStatus status;
 	private LocalDateTime from;
 	private LocalDateTime to;
 
-	public Specification<OrderEntity> getSpecificationForArtist(Long shopId) {
-		return getSpecification(root -> root.get("shop").get("id"), shopId);
-	}
-
-	public Specification<OrderEntity> getSpecificationForUser(Long userId) {
-		return getSpecification(root -> root.get("orderGroup").get("user").get("id"), userId);
-	}
-
-	private Specification<OrderEntity> getSpecification(Function<Root<OrderEntity>, Path<Long>> identifier, Long id) {
+	public Specification<OrderGroupEntity> getSpecificationForUser(Long userId) {
 		return (root, cQuery, builder) -> {
 			List<Predicate> predicates = new ArrayList<>();
-			predicates.add(builder.equal(identifier.apply(root), id));
+			predicates.add(builder.equal(root.get("user").get("id"), userId));
 			if (status != null) {
-				predicates.add(builder.equal(root.get("status"), status.getByteValue()));
+				Join<OrderEntity, OrderGroupEntity> orderJoin = root.join("orders");
+				predicates.add(builder.equal(orderJoin.get("status"), status.getByteValue()));
 			}
 			if (from != null) {
 				predicates.add(builder.greaterThanOrEqualTo(root.get("createdDate"), from));
 			}
 			if (to != null) {
-				predicates.add(builder.lessThanOrEqualTo(root.get("createdDate"), to));
+				predicates.add(builder.lessThanOrEqualTo(root.get("createdDate"), from));
 			}
 			return builder.and(predicates.toArray(new Predicate[0]));
 		};
 	}
-
-
 }
