+++ lib_artiexh-model/src/main/java/com/artiexh/model/rest/order/response/CheckoutResponse.java
@@ -1,11 +1,15 @@
 package com.artiexh.model.rest.order.response;
 
+import com.artiexh.model.domain.PaymentMethod;
 import com.artiexh.model.domain.UserAddress;
 import com.artiexh.model.rest.transaction.OrderTransactionResponse;
 import com.artiexh.model.rest.user.UserOrderResponse;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import com.fasterxml.jackson.databind.ser.std.ToStringSerializer;
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 import java.util.Set;
 
@@ -19,4 +23,5 @@ public class CheckoutResponse {
 	private Set<UserOrderResponse> orders;
 	private UserAddress shippingAddress;
 	private OrderTransactionResponse currentTransaction;
+	private PaymentMethod paymentMethod;
 }
